+++ History.md
@@ -7,6 +7,7 @@
   * `Puma.stats` now returns a Hash instead of a JSON string (#2086)
   * `GC.compact` is called before fork if available (#2093)
   * Add `requests_count` to workers stats. (#2106)
+  * Faster phased restart and worker timeout
 
 * Bugfixes
   * Your bugfix goes here (#Github Number)